\name{internal-functions}
\alias{emMixGGM}
\alias{graphPenalty}
\alias{searchGGMGA}
\alias{fitnessGGM}
\alias{searchGGMStepwise}
\alias{startStepwise}
\alias{estepmggm}
\alias{findspouse}
\alias{icf}
\alias{conggm}
\alias{profileloglik}
\alias{zcalc}

\title{
Internal \code{mggm} function
}
\description{
Internal function not to be called by the user.
}

\keyword{internal}
