\name{mixggm-package}
\alias{mixggm-package}
\alias{mixggm}
\docType{package}

\title{Mixtures of Gaussian graphical models}

\description{Mixtures of Gaussian graphical models for model-based clustering with sparse covariance and concentration matrices.}

\details{A package implementing mixtures of Gaussian graphical models for model-based clustering with sparse covariance and concentration matrices. Model fitting is carried out by means of a structural-EM algorithm for parameter estimation and graph structure search.

The main functions are \code{\link{mixGGM}}, \code{\link{searchGGM}} and \code{\link{fitGGM}}. }

\author{
Michael Fop, Luca Scrucca and Thomas Brendan Murphy.

Maintainer: Michael Fop \email{michael.fop@ucd.ie}
}

\references{
Fop, M., Murphy, T.B., and Scrucca, L. (2018). Model-based clustering with sparse covariance matrices. \emph{Statistics and Computing}. To appear.
}

\keyword{package}
