% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixGGM.R
\name{MixGGM_function}
\alias{MixGGM_function}
\title{MixGGM function}
\usage{
MixGGM_function(X, K, penalty, init)
}
\arguments{
\item{K}{number of components}

\item{penalty}{"no_sparsity", "CV", "StARS" , "fixed" , "iterative"}

\item{init}{"Kmeans"}

\item{file}{name of the csv file as input data}

\item{out}{"precision" , "partial" , "adj"}

\item{threshold}{a value between 0 and 1 which indicates the threshod to generate the adjacency matrix}
}
\description{
This function receives the sample-taxa count matrix and cluser the samples based on the MixGGM method.
}
